+++ CHANGES.md
@@ -1,3 +1,17 @@
+Version 1.1
+-----------
+* Added module `Language.Haskell.TH.Desugar.Expand`, which allows for expansion
+  of type synonyms in desugared types.
+* Added `Show`, `Typeable`, and `Data` instances to desugared types.
+* Fixed bug where an as-pattern in a `let` statement was scoped incorrectly.
+* Changed signature of `dsPat` to be more specific to as-patterns; this allowed
+  for fixing the `let` scoping bug.
+* Created new functions `dsPatOverExp` and `dsPatsOverExp` to allow for easy
+  desugaring of patterns.
+* Changed signature of `dsLetDec` to return a list of `DLetDec`s.
+* Added `dsLetDecs` for convenience. Now, instead
+  of using `mapM dsLetDec`, you should use `dsLetDecs`.
+
 Version 1.0
 -----------